Revert "refactor `scryptsy` into jellyfish-wallet-encrypted/test (#516)"
This reverts commit 053839f211d7d674ebd8c291c7c140962176dd27.
