+++ packages/jellyfish-wallet-encrypted/package.json
@@ -35,10 +35,10 @@
     "build": "tsc -b ./tsconfig.build.json"
   },
   "dependencies": {
-    "@defichain/jellyfish-wallet-mnemonic": "0.0.0"
+    "@defichain/jellyfish-wallet-mnemonic": "0.0.0",
+    "scryptsy": "^2.1.0"
   },
   "devDependencies": {
-    "scryptsy": "^2.1.0",
     "@types/scryptsy": "^2.0.0"
   }
 }
