+++ packages/jellyfish-wallet-encrypted/src/index.ts
@@ -1,3 +1,4 @@
 export * from './hd_node'
 export * from './provider/scrypt_provider'
+export * from './provider/scrypt_simple'
 export * from './scrypt'
